[2] Added an even more optimized algorithm
This one, called `evaluate_string`, directly iterates over the
input string slice. It's sort of a state machine. It takes
different actions based on whether it encounters a number, a space,
or a newline in the string. No heap allocations whatsoever.
A perfectly linear time solution. Arguably, the code is a bit
difficult to understand, but hey, it's fast!